# internal/ports/httpport/openapi_types.gen.go
// Package httpport provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package httpport

import (
	"time"
)

// Authenticated defines model for Authenticated.
type Authenticated struct {
	AccessToken string `json:"accessToken"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// PostLogin defines model for PostLogin.
type PostLogin struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// PostRegister defines model for PostRegister.
type PostRegister struct {
	Email    string `json:"email"`
	Password string `json:"password"`
	Uuid     string `json:"uuid"`
}

// User defines model for User.
type User struct {
	CreatedAt time.Time `json:"createdAt"`
	Email     string    `json:"email"`
	UpdatedAt time.Time `json:"updatedAt"`
	Uuid      string    `json:"uuid"`
}

// LoginUserJSONRequestBody defines body for LoginUser for application/json ContentType.
type LoginUserJSONRequestBody = PostLogin

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = PostRegister